tests/__init__.py → scrapy_poet/utils/testing.py
