+++ scrapy_poet/utils/testing.py
@@ -3,11 +3,12 @@ from inspect import isasyncgenfunction
 from typing import Dict
 from unittest import mock
 
-from scrapy import signals
+from scrapy import Spider, signals
 from scrapy.crawler import Crawler
 from scrapy.exceptions import CloseSpider
 from scrapy.settings import Settings
 from scrapy.utils.python import to_bytes
+from scrapy.utils.test import get_crawler as _get_crawler
 from twisted.internet import reactor
 from twisted.internet.defer import inlineCallbacks
 from twisted.internet.task import deferLater
@@ -151,6 +152,10 @@ def crawl_single_item(
     return item, url, crawler
 
 
+def get_download_handler(crawler, schema):
+    return crawler.engine.downloader.handlers._get_handler(schema)
+
+
 def make_crawler(spider_cls, settings):
     if not getattr(spider_cls, "name", None):
 
@@ -161,6 +166,33 @@ def make_crawler(spider_cls, settings):
         Spider.__module__ = spider_cls.__module__
         spider_cls = Spider
     return Crawler(spider_cls, settings)
+
+
+def setup_crawler_engine(crawler: Crawler):
+    """Run the crawl steps until engine setup, so that crawler.engine is not
+    None.
+    https://github.com/scrapy/scrapy/blob/8fbebfa943c3352f5ba49f46531a6ccdd0b52b60/scrapy/crawler.py#L116-L122
+    """
+
+    crawler.crawling = True
+    crawler.spider = crawler._create_spider()
+    crawler.engine = crawler._create_engine()
+
+    handler = get_download_handler(crawler, "https")
+    if hasattr(handler, "engine_started"):
+        handler.engine_started()
+
+
+class DummySpider(Spider):
+    name = "dummy"
+
+
+def get_crawler(settings=None, spider_cls=DummySpider, setup_engine=True):
+    settings = settings or {}
+    crawler = _get_crawler(settings_dict=settings, spidercls=spider_cls)
+    if setup_engine:
+        setup_crawler_engine(crawler)
+    return crawler
 
 
 class CollectorPipeline:
